Separate functionality to util
