+++ flask_util.py
@@ -1,22 +1,13 @@
 from pathlib import Path
 from typing import Dict
-from uuid import uuid4
 from datetime import datetime
+from db_model import Session, Upload, User
 import os
 import json
 
+
 UPLOADS_FOLDER = "uploads"
 OUTPUTS_FOLDER = "outputs"
-
-
-def generate_uid() -> str:
-    """
-    Generates a unique identifier (UID) using the UUID4 algorithm.
-
-    Returns:
-        str: The generated unique identifier.
-    """
-    return str(uuid4())
 
 
 def get_output_path(filename: str) -> str:
@@ -78,3 +69,56 @@ def set_path():
     """
     Path(UPLOADS_FOLDER).mkdir(parents=True, exist_ok=True)
     Path(OUTPUTS_FOLDER).mkdir(parents=True, exist_ok=True)
+
+
+def save_upload(file):
+    """
+    Saves the uploaded file as an anonymous upload.
+
+    This function creates an Upload object in the database to represent the uploaded file
+    and saves the file to the uploads folder using a generated UID. The function returns
+    the UID associated with the uploaded file.
+
+    Args:
+        file (FileStorage): The uploaded file to be saved.
+
+    Returns:
+        str: The UID associated with the uploaded file.
+    """
+    with Session() as session:
+        anonymous_upload = Upload(filename=file.filename, upload_time=datetime.now())
+        session.add(anonymous_upload)
+        session.commit()
+        _, file_type = os.path.splitext(file.filename)
+        file.save(os.path.join(UPLOADS_FOLDER, f"{anonymous_upload.uid}{file_type}"))
+        return anonymous_upload.uid
+
+
+def save_upload_with_user(file, email: str):
+    """
+    Saves the uploaded file with the associated user.
+
+    This function creates a User object in the database if the user with the provided
+    email doesn't exist. It then creates an Upload object associated with the user and
+    saves the file to the uploads folder using a generated UID. The function returns
+    the UID associated with the uploaded file.
+
+    Args:
+        file (FileStorage): The uploaded file to be saved.
+        email (str): The email of the user associated with the uploaded file.
+
+    Returns:
+        str: The UID associated with the uploaded file.
+    """
+    with Session() as session:
+        user = session.query(User).filter_by(email=email).first()
+        if not user:
+            user = User(email=email)
+            session.add(user)
+            session.commit()
+        user_upload = Upload(filename=file.filename, upload_time=datetime.now(), user=user)
+        session.add(user_upload)
+        session.commit()
+        _, file_type = os.path.splitext(file.filename)
+        file.save(os.path.join(UPLOADS_FOLDER, f"{user_upload.uid}{file_type}"))
+        return user_upload.uid
